Fix nested nginx location blocks
The nested blocks don't inherit the conditional rewrite logic, so it
needs to be duplicated. If it is missing, deploy attempts to
clojars.org/repo will 403 when trying to read the metadata.

Deploys to repo.clojars.org will work, since the metadata is read from
the CDN instead of being redirected there.